fix(cluster): resolve GCC 11 -Werror build failures in CI
- Remove unused AERON_ARCHIVE_CONTROL_SESSION_CLOSED_MSG variable
- Remove unused replication_session_track_action function
- Fix sign-compare in recording_session min_size call
- Fix snprintf truncation in delete_segments error message
- Remove duplicate _DEFAULT_SOURCE define in mark_file

+++ aeron-archive/src/main/c/server/aeron_archive_recording_session.c
@@ -202,8 +202,8 @@ int aeron_archive_recording_session_create(
     _session->progress_event_position = image_constants.join_position;
     _session->join_position = image_constants.join_position;
     _session->block_length_limit = min_size(
-        image_constants.term_buffer_length,
-        (size_t)(file_io_max_length > 0 ? file_io_max_length : image_constants.term_buffer_length));
+        (size_t)image_constants.term_buffer_length,
+        file_io_max_length > 0 ? (size_t)file_io_max_length : (size_t)image_constants.term_buffer_length);
 
     /* Create the recording writer */
     if (aeron_archive_recording_writer_create(
